backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
pydantic==2.6.1
groq==0.4.2
pillow==10.2.0
numpy==1.26.4
python-magic==0.4.27
requests==2.31.0 